AWS_Autoscale/lambda_code/autoscale/autoscale/requirements.txt
argcomplete==1.9.3
asn1crypto==0.24.0
boto3==1.9.140
botocore==1.12.140
certifi==2019.3.9
cffi==1.12.3
cfn-flip==1.2.0
chardet==3.0.4
Click==7.0
cryptography==2.6.1
Django==2.2.1
docutils==0.14
durationpy==0.5
future==0.16.0
hjson==3.0.1
idna==2.8
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
pem==19.1.0
placebo==0.9.0
pycparser==2.19
pyOpenSSL==19.0.0
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2019.1
PyYAML==5.1
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.4.6
Unidecode==1.0.23
urllib3==1.25.2
Werkzeug==0.15.2
wsgi-request-logger==0.4.6
zappa==0.48.2
